Bug Fix: Refactor class `Entity`, `Environment`, `Resource`
* Refactor `start/start_in_pool/stop/stop_in_pool` of `Environment`
  class to make code clean.
* Fix a bug that int `Environment.start` the `_wait_started` is called
  twice if a resource's `async_start` attribute is False.
* Fix a bug that in `Environment.start_in_pool` a resource will be
  started even its `auto_start` attribute is False.
* No need to maintain 2 copies of `start/stop` methods of `Resource`
  and its derive class `Driver`, the only difference is that there are
  customized `pre_start/post_start/pre_stop/post_stop` to be called.
  The `Environment` class works for `Driver` but it should be adapted
  to `Resource` (e.g. `resource.wait` is called twice unexpectedly).
* While start or stop resources, should check if it's already aborted,
  or the status transition will not be correct and raise exception.
* If exception raised during `stop`, the resource status should also
  be updated to avoid possible issues (In interactive base module we
  have some code to deal with the status that could be removed).
* Move `post_start` and `post_stop` methods out of wait handlers, it's
  confusing that `pre_start` and `post_start` are called in `start`,
  while the others are called in a protected methods, hard to maintain.
* `Worker`s are maintained by `Pool`, so no need to change status to
  `STARTED` in its `staring` method, in this time worker is not fully
   initialized and the heartbeat check thread might unexpectedly exit.
* `_log_exception` always places error messages into `start_exceptions`
  but for `stop_in_pool` error messages should go to `stop_exceptions`.
* In debug log use '%s' format rather than '%r' because unique name is
  fairly enough, better than "CLASS_NAME <object at 0x00...>".
* Just like `start` and `stop`, it's better log basic debug info in
  `abort` method rather than in `aborting` method everywhere.
* Fix wrong log in `_abort_entity` method.
* `__eq__` method is already added for `ResourceStatus` so we can use
  `res.status == ST` instead of `res.status.tag == ST` everywhere.
* Remove lint error on try...finally... statements.
* Update testcases.

+++ testplan/runners/pools/connection.py
@@ -277,12 +277,11 @@ class Server(entity.Resource, metaclass=abc.ABCMeta):
 
     def starting(self):
         """Server starting logic."""
-        self.status.change(self.status.STARTED)
+        self.status.change(self.status.STARTED)  # Start is async
 
     def stopping(self):
         """Server stopping logic."""
-
-        self.status.change(self.status.STOPPED)
+        self.status.change(self.status.STOPPED)  # Stop is async
 
     def aborting(self):
         """Abort policy - no abort actions are required in the base class."""
@@ -295,10 +294,10 @@ class Server(entity.Resource, metaclass=abc.ABCMeta):
         connection manager is started and will be automatically unregistered
         when it is stopped.
         """
-        if self.status.tag != self.status.STARTED:
+        if self.status != self.status.STARTED:
             raise RuntimeError(
-                "Can only register workers when started. Current state is "
-                "{}".format(self.status.tag)
+                "Can only register workers when started."
+                f" Current state is {self.status.tag}."
             )
 
     @abc.abstractmethod
@@ -396,10 +395,12 @@ class ZMQServer(Server):
     def _close(self):
         """Closes TCP connections managed by this object.."""
         self.logger.debug("Closing TCP connections for %s", self.parent)
-        self._sock.close()
-        self._sock = None
-        self._zmq_context.destroy()
-        self._zmq_context = None
+        if self._sock is not None:
+            self._sock.close()
+            self._sock = None
+        if self._zmq_context is not None:
+            self._zmq_context.destroy()
+            self._zmq_context = None
         self._address = None
 
     def stopping(self):
